+++ src/main/java/lesson_e/Main.java
@@ -1,7 +1,5 @@
 package lesson_e;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
 import java.util.concurrent.Semaphore;
@@ -32,129 +30,7 @@ public class Main {
     }
 }
 
-class Car implements Runnable {
-    private static int CARS_COUNT;
-    private CyclicBarrier barrier;
-    private Object monitor;
-    private Race race;
-    private int speed;
-    private String name;
-
-    static {
-        CARS_COUNT = 0;
-    }
 
 
-    public String getName() {
-        return name;
-    }
 
-    public int getSpeed() {
-        return speed;
-    }
 
-    public Car(Object monitor, CyclicBarrier barrier, Race race, int speed) {
-        this.race = race;
-        this.speed = speed;
-        CARS_COUNT++;
-        this.name = "Участник #" + CARS_COUNT;
-        this.barrier = barrier;
-        this.monitor = monitor;
-    }
-
-    @Override
-    public void run() {
-        try {
-            System.out.println(this.name + " готовится");
-            Thread.sleep(500 + (int) (Math.random() * 800));
-            System.out.println(this.name + " готов");
-            barrier.await();
-            barrier.await();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        for (int i = 0; i < race.getStages().size(); i++) {
-            race.getStages().get(i).go(this);
-        }
-
-        //проверка победителя
-        synchronized (monitor) {
-            if (Main.WIN_COUNT > 0) {
-                System.out.println(this.name + " WIN " + " занял " + ++Main.WIN + " место.");
-                Main.WIN_COUNT--; //уменьшпем места победителей
-            }
-        }
-        try {
-            barrier.await(); //ждем пока гонка закончится
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        } catch (BrokenBarrierException e) {
-            e.printStackTrace();
-        }
-    }
-
-}
-
-class Race {
-    private ArrayList<Stage> stages;
-
-    public ArrayList<Stage> getStages() {
-        return stages;
-    }
-
-    public Race(Stage... stages) {
-        this.stages = new ArrayList<>(Arrays.asList(stages));
-    }
-}
-class Road extends Stage {
-    public Road(int length) {
-        this.length = length;
-        this.description = "Дорога " + length + " метров";
-    }
-
-    @Override
-    public void go(Car c) {
-        try {
-            System.out.println(c.getName() + " начал этап: " + description);
-            Thread.sleep(length / c.getSpeed() * 1000);
-            System.out.println(c.getName() + " закончил этап: " + description);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    }
-}
-abstract class Stage {
-    protected int length;
-    protected String description;
-
-    public String getDescription() {
-        return description;
-    }
-
-    public abstract void go(Car c);
-}
-class Tunnel extends Stage {
-    public Tunnel() {
-        this.length = 80;
-        this.description = "Тоннель " + length + " метров";
-    }
-
-    @Override
-    public void go(Car c) {
-        try {
-            try {
-                System.out.println(c.getName() + " готовится к этапу(ждет): " + description);
-                Main.SEMAPHORE_TUNNEL.acquire();
-                System.out.println(c.getName() + " начал этап: " + description);
-                Thread.sleep(length / c.getSpeed() * 1000);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            } finally {
-                System.out.println(c.getName() + " закончил этап: " + description);
-                Main.SEMAPHORE_TUNNEL.release();
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-}
