+++ src/main/java/lesson_e/Race.java
@@ -1,0 +1,33 @@
+package lesson_e;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+
+class Race {
+    private ArrayList<Stage> stages;
+
+    public ArrayList<Stage> getStages() {
+        return stages;
+    }
+
+    public Race(Stage... stages) {
+        this.stages = new ArrayList<>(Arrays.asList(stages));
+    }
+}
+class Road extends Stage {
+    public Road(int length) {
+        this.length = length;
+        this.description = "Дорога " + length + " метров";
+    }
+
+    @Override
+    public void go(Car c) {
+        try {
+            System.out.println(c.getName() + " начал этап: " + description);
+            Thread.sleep(length / c.getSpeed() * 1000);
+            System.out.println(c.getName() + " закончил этап: " + description);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+}
