+++ src/main/java/lesson_e/Stage.java
@@ -1,0 +1,12 @@
+package lesson_e;
+
+abstract class Stage {
+    protected int length;
+    protected String description;
+
+    public String getDescription() {
+        return description;
+    }
+
+    public abstract void go(Car c);
+}
