+++ src/main/java/lesson_e/Tunnel.java
@@ -1,0 +1,27 @@
+package lesson_e;
+
+class Tunnel extends Stage {
+    public Tunnel() {
+        this.length = 80;
+        this.description = "Тоннель " + length + " метров";
+    }
+
+    @Override
+    public void go(Car c) {
+        try {
+            try {
+                System.out.println(c.getName() + " готовится к этапу(ждет): " + description);
+                Main.SEMAPHORE_TUNNEL.acquire();
+                System.out.println(c.getName() + " начал этап: " + description);
+                Thread.sleep(length / c.getSpeed() * 1000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            } finally {
+                System.out.println(c.getName() + " закончил этап: " + description);
+                Main.SEMAPHORE_TUNNEL.release();
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+}
